+++ backend/fittoring/src/test/java/fittoring/mentoring/business/repository/MemberRepositoryTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import fittoring.mentoring.business.model.Member;
 import fittoring.mentoring.business.model.Phone;
 import fittoring.mentoring.business.model.password.Password;
-import fittoring.mentoring.business.repository.helper.MentoringPaginationHelper;
+import fittoring.mentoring.infra.querydsl.helper.MentoringPaginationHelper;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
