+++ backend/fittoring/src/test/java/fittoring/mentoring/business/service/PhoneVerificationServiceTest.java
@@ -6,7 +6,7 @@ import fittoring.mentoring.business.exception.InvalidPhoneVerificationException;
 import fittoring.mentoring.business.model.Phone;
 import fittoring.mentoring.business.model.PhoneVerification;
 import fittoring.mentoring.business.repository.PhoneVerificationRepository;
-import fittoring.mentoring.business.repository.helper.MentoringPaginationHelper;
+import fittoring.mentoring.infra.querydsl.helper.MentoringPaginationHelper;
 import fittoring.mentoring.infra.CodeGeneratorStub;
 import fittoring.mentoring.presentation.dto.VerificationCodeRequest;
 import fittoring.util.DbCleaner;
